Optimized memcpy to use SSE for x86_64, added optimized strlen ( also using SSE, for x86_64 )

+++ inc/stddef.h
@@ -4,7 +4,7 @@
 
 #define NULL (void*)0
 
-typedef unsigned int size_t;
+typedef unsigned long uintptr_t;
 
 /*
  * Take advantage  of GCC builtins, as  the  latter  method, despite
